+++ trunk/hello_world/objects/world.c
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-int calculate_world( obj_loc_t obj_loc, int id )
+int calculate_world( obj_loc_t obj_loc, long id )
 {
     // Does nothing.
